+++ lib/minglib/gui/label.hpp
@@ -5,8 +5,8 @@
 
 class Label : public GUIObject {
 public:
-    Label(long id, const char *text, int x, int y, int width, int height);
-    Label(Window *parent, long id, const char *text, int x, int y, int width, int height);
+    Label(long id, const char *text, UINT16 x, UINT16 y, UINT16 width, UINT16 height);
+    Label(Window *parent, long id, const char *text, UINT16 x, UINT16 y, UINT16 width, UINT16 height);
 };
 
 #endif
